+++ log4j-api/src/main/java/org/apache/logging/log4j/util/DummyRecyclerFactory.java
@@ -14,7 +14,7 @@
  * See the license for the specific language governing permissions and
  * limitations under the license.
  */
-package org.apache.logging.log4j.layout.template.json.util;
+package org.apache.logging.log4j.util;
 
 import java.util.function.Consumer;
 import java.util.function.Supplier;
