+++ log4j-api/src/main/java/org/apache/logging/log4j/util/QueueingRecyclerFactory.java
@@ -14,27 +14,27 @@
  * See the license for the specific language governing permissions and
  * limitations under the license.
  */
-package org.apache.logging.log4j.layout.template.json.util;
+package org.apache.logging.log4j.util;
 
+import java.util.Queue;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-public class ThreadLocalRecyclerFactory implements RecyclerFactory {
+public class QueueingRecyclerFactory implements RecyclerFactory {
 
-    private static final ThreadLocalRecyclerFactory INSTANCE =
-            new ThreadLocalRecyclerFactory();
+    private final Supplier<Queue<Object>> queueSupplier;
 
-    private ThreadLocalRecyclerFactory() {}
-
-    public static ThreadLocalRecyclerFactory getInstance() {
-        return INSTANCE;
+    public QueueingRecyclerFactory(final Supplier<Queue<Object>> queueSupplier) {
+        this.queueSupplier = queueSupplier;
     }
 
     @Override
     public <V> Recycler<V> create(
             final Supplier<V> supplier,
             final Consumer<V> cleaner) {
-        return new ThreadLocalRecycler<>(supplier, cleaner);
+        @SuppressWarnings("unchecked")
+        final Queue<V> queue = (Queue<V>) queueSupplier.get();
+        return new QueueingRecycler<>(supplier, cleaner, queue);
     }
 
 }
