+++ log4j-api/src/main/java/org/apache/logging/log4j/util/StringParameterParser.java
@@ -14,9 +14,7 @@
  * See the license for the specific language governing permissions and
  * limitations under the license.
  */
-package org.apache.logging.log4j.layout.template.json.util;
-
-import org.apache.logging.log4j.util.Strings;
+package org.apache.logging.log4j.util;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -24,6 +22,8 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.Callable;
+
+import org.apache.logging.log4j.util.Strings;
 
 public final class StringParameterParser {
